chore: refactor / fix linting
